Updated the refactor code
