+++ inc/Interfaces/Kernel.php
@@ -14,7 +14,7 @@ interface Kernel {
 	/**
 	 * Register logic.
 	 *
-	 * @since 1.0.0
+	 * @since 1.2.0
 	 *
 	 * @return void
 	 */
